Increase Commons linkpurge rate limit for patrollers
Linkpurge rate is increased on null edits which can happen frequently
during mass automated edits, and it prevents further edits once the
threshold is reached. As a workaround, raise the limit for patrollers
(who also have edit rate limit exceptions and as such are the most
likely to run into this).

Bug: T245214
Change-Id: Id8ca31644b6075b4c83f12c84db09e83f41a5a7c