Make Beta labs CSP settings be same as prod but with beta urls
Beta should look as much as prod as possible, so lets make the
CSP settings be the same.

Note: This means that only beta urls are in beta CSP, so you
will get a warning if you try to include a gadget from prod
at beta. I think that's the sane thing to do.

Bug: T245983
Change-Id: Ia04e96de7da49f00a1868b8427a76bd5107d0039