db-eqiad,db-codfw.php: Add es4 as new ES, for initial testing
es4 is a new external store section.
Per T246072#5916506, the idea would be to add es4 on
$templateOverridesByCluster so we can manually test it.
Once we are happy with it, on a different patch, we can
add it to $wgDefaultExternalStore so writes (and reads)
will be enabled everywhere.

Same steps would be repeated for es5 later as well.

Bug: T246072
Change-Id: I4ba5944789e91540089cbb4c67e03676f3eb5b65