Parsoid: Use the version of Parsoid in $IP/vendor
This is a great deal simpler than the current system, and allows us
to drop the Parsoid-deploy repo entirely, if we wish to, whilst the
ability to ship new versions of Parsoid is retained, albeit reduced
in rate to (roughly) once a week.

This means that Parsoid versions will differ for different versions
of MediaWiki (so group0 will get "new" Parsoid and group1 and 2 the
"old" Parsoid for immediate testing and a go/no-go decision is made
when each new train initially goes live).

Scandium special-snowflake status may not work with this change; to
be investigated before and after deployment.

Bug: T240055
Change-Id: Ia9026b1da57d1e4397a946cce50b708b0e953c62