Do not update the globals cache file while opcache needs regeneration
We use the mtime of InitialiseSettings for invalidating the static
cache of globals, but given we refresh the content of the files only
every opcache.revalidate_freq seconds, this leads to a race condition
where old values are written to cache, and never recover.

To fix this issue, refuse to refresh the cache file until the grace
period has expired. This way, we avoid the race condition and still
get the new code as soon as the opcache is refreshed.

I tried to keep the patch as small as possible, but alternative
approach could be to check the staleness of the opcache data for
InitialiseSettings.php  using opcache_get_status( true ), but this is
quite expensive in terms of memory consumption and probably in terms
of computing time too.

Bug: T236104
Change-Id: I9bb757f2c883c718b978d38c811d861843f73696