Move Defines.php mock from tests/ to src/
Looks like this is the only run-time relation between
InitialiseSettings.php and MediaWiki core that we can't (easily)
shake so let's accept this as source code for now and make it
runnable both in and outside MW context with a conditional.

In the future we might make this unconditional once we can guruantee
we'll never have to invoke the config builder from the web, but for
now let's leave our options open.

Change-Id: Idc26716abef5bff001450083835b7b0ea254d72b