multiversion: Remove support for passing file path to readDbListFile()
The following kinds of calls are no longer supported:

* MWWikiversions::readDbListFile('foo.dblist')

This was used in two places only (both in wmf-config/import.php),
which this commmit fixes.

* MWWikiversions::readDbListFile('/usr/local/mediawiki/dblists/foo.dblist')

This seems to have worked by accident, in that basename() not only
strips ".dblist" but also strips the directory path and ignores it
completely.

This kind of call was only seen in one place (mediawiki/extensions/SecurePoll)
in an import script used once in 2017 (presumably unused), but fixed nonetheless
in Icd81da8d5315bbb728.

Change-Id: I9d658ff41b78917fad3da333de490df13e47f364