tests: Reduce 'family' assertion to just 'wiki-suffix disambig'
== The free config tag

The dblists exists mainly so that wmf-config reads each of these files,
sees which one contain the current wiki, and use the dblist name as the
'tags' for expanding the InitialiseSettings.php. This logic lives in
MWConfigCacheGenerator::getMWConfigForCacheing and SiteConfiguration.php.

The thing is, every wiki already gets one such tag for free, based on its
dbname suffix. This is extracted in SiteConfiguration based on wgConf->suffixes,
where every dbname is split into $lang and $site, where $site (aka $dbSuffix)
is passed down as one of the tags.

Right now when running on nlwiktionary, we read all of wikinews.dblist,
wiktionary.dblist, wikivoyage.dblist etc. just to find that it is one
wiktionary.dblist and then add 'wiktionary' to the list of tags.
We already know this based on $site/$dbSuffix. The logic for this is already
in-place and has been working for years. The tags are simply array_unique'ed
right now, thus effectively ignoring most of these dblists.

The exception is "wikipedia" which we still need as dblist. This is because
those wikis share the same suffix as special wikis (namely "wiki"). For this
reason, we don't register "wikipedia" as a suffix. Instead the "wikipedia"
$site, is casted as $dbSuffix "wiki" in getMWConfigForCacheing.

(If this casting were to not exist, then all special wikis would be getting
configurations for "wikipedia" applied).

== Testing

The unit test to enforce a 1:1 mapping between wiki IDs and family dblists
is redundant, because the fact that each wiki ID is split between $lang
and $site suffix, means there is naturally only one suffix anyway.

The main thing we do want to assert still is that within the '*wiki' suffix
we only have the wiki in one of the related dblists (in this case: wikipedia,
wikimedia-chapter, and special).

Bug: T169821
Change-Id: If8aa7504dc529504ebd809e193961c42ca5d8cb4