MWConfigCacheGenerator: Stop reading most wiki-family dblist files
See commit message of If8aa7504dc529 for details.

In short: We don't need to read all these files to figure out
which one the current wiki is in, because the dbname's suffix
already tells us that information, which we already extract
and already provide to $wgConf->getAll() as extra tag.

This reduces the number of dblist files we need to read at run-time
by about ~20% (8/43).

Bug: T169821
Change-Id: Ib2aaf6540d85e2ef1b0b77e088fea8bd2e8eb9bb