multiversion: Update copy of SiteConfiguration to match current MW
This splits the configuration interpreter from the cache generating
and YAML logic to make the class easier to update in the future,
as it's basically just a straight copy now.

In the next commit I'll also remove the tests/SiteConfiguration.php
copy that we have and make the tests use StaticSiteConfiguration.

Change-Id: Ic344b48a1f88388efd58e5dad15d4f176d5a2b97