tests: Re-enable 'family' dblist test
Follows-up c74a33a38. I removed this thinking that we'd remove
the dblists as they'd be obsolete for wmf-config.

While that is true, they are still useful for use with
'foreachwikiindblist' and puppetised maintenance jobs.

As such, keeping the dblist files still makes sense, and we
can continue asserting that each wiki is only in one of them.

Note that to prevent breaking that family == suffix group,
the testDatabaseSuffixMatchProject() test also still exists,
which continues to do its job. (I've simplified it a bit here)

Bug: T239301
Change-Id: I9749e571878ead7ab808b9ca6f5f5757af548b2d