Resolve 'wgEnableUploads' ambiguity for wikinews wikis
The 'commonsuploads' and 'wikinews' wikis overlap by 8 wikis, which creates
an ambiguity. Currently, the 'wikinews' list seems to apply first
(given the load order in MWConfigCacheGenerator::$dbLists).

7 of the overlapping wikis actually have their setting explicitly set,
which leaves 1 wikinews wiki, liwikinews, as the only one for its
its config is truly ambiguous/fragile.

Fix this by removing the overlapping 'wikinews' override, and list
the ones for which wgEnableUploads should be false explicitly.

Also remove the ones for which it is set to true, given that the
default value of true is no longer blocked behind the 'wikinews'
override.

Change-Id: I52bb702438403d81aed73971a52378e894844a9d