tests: Remove outdated 'testExpressionListsNaming' test
The testNoExpressionListUsedInSettings() is already asserting
what this one does. The only diffence is that testExpressionListsNaming()
was also harcoding all lazy lists that we have in existence ('echo' and 'open')
which doesn't seem particularly useful as those can be created
without concern so long they are not used on the web, no need to
require the test being updated each time.

This test was originally to assert that lazy lists are named as
"-computed.dblist", but we changed that last year at which point
the test became obsolete.

Change-Id: I67f657b3d1f5b6a85fbb9ae7fe9e347f2d850059