multiversion: Make buildDBLists.php both create and delete dblist files
If no wikis are left in a dblist, it currently leaves the file behind
and containing the last set of wikis that used to be in it before the
commit that removed them, this because the script doesn't know that it
used to exist.

As quick solution for now, scan the directory afterwards and delete any
files that are untracked (with a few exceptions).

Bug: T223602
Change-Id: I8a5f28bbbb2144f13a32e83ec9a64d1ef09ea8d3