dblists: Remove 'labtestwiki' dblist containing 'labtestwiki'
This list is a problem in two ways:

1. The dblist file was untracked because config/labtestwiki.yaml
   did not specify that it was in a tag called 'labtestwiki'.

   When pruning the directory and running 'composer buildDBLists'
   this file goes missing.

   Once this list is removed, I8a5f28bbbb214 is unblocked
   which will make sure this problem won't resurface.

2. Due to 'labtestwiki' being a tag, it can clash with other tags
   that it overlaps with.

   For example, the labtestwiki wiki is part of tag 'wikitech' and
   tag 'labtestwiki'. $wgSessionCacheType is set to memc for wikitech
   to kask for labtestwiki. It's undefined which one should win
   from a tagging perspective.

   This and other tag-level ambiguities will be caught by CI
   going forward once this lands, as part of Ia53907882a23d2.

Removal has no impact on wmf-config processing because it was
named after the one wiki it contained, which means that
any setting for 'labtestwiki' still applies just the same.

Also, it wasn't loaded in MWConfigCacheGenerator for use
at run-time in web requests, which means reliance on that
would have failed regardless.

The list was originally created to be the database section list
for "lab test". It has since been renamed to s11 for that purpose,
but this file was left behind.

Bug: T223602
Change-Id: I22a3c2a82f7be4a582302d79dea02d19ab611fdd