dblists: Remove closed wikis from commonsuploads (angwikibooks, iewikibooks)
The $wgGroupPermissions config (via wmf-config 'groupOverrides') is set
for both 'closed' and 'commonsuploads' to different values, which
creates an ambiguity as to whether which one wins.

Very very fortunately, the internal order in MWConfigCacheGenerator::$dbLists
enumerates 'closed' before 'commonsuploads' which means the settings for
'closed' won in this case, and so these two wikis in fact were not
accidentally re-opened when the 'commonsuploads' setting was added to it.

This conflict was detected via the unit test from Ia53907882a23d2a16.

Effective change to config settings that use this dblist:

* $wgGroupPermissions: None, 'closed' was already winning.
* $wgEnableUploads: None, 'commonsuploads' and 'default both set true.
* $wgUploadNavigationUrl: The setting for these two closed wikis to
  make "Upload file" point to Commons will no longer apply.
  This seems fine given the wiki is closed and no longer accepts
  contributions anyhow.

Change-Id: Idb69b82f56c85432b4430d4a02e9a29fbbcf7d3a