multiversion: Structure buildDBLists.php to prevent labs info in prod
The fragmentation in knownDBLists by realm suggested that they
are stored separately, but that is not (yet?) the case. I couldn't
find a way to break out of this as the inline checks correctly
prevented this. But, structuring it in a way where they are stored
in one array and written once seems easier to extend in the future.

It also removes a risk where the second iteration (labs) could maybe
somehow "win" and overwrite any clashes with labs-specific info.
This isn't possible right now, but makes it harder to get wrong
in the future.

With knownDBLists as one array, the processing and writing steps
no longer overlap, which means the hacks I added in 41c24b38731a17
can be removed in favour of a clean 'rm *.dblist' followed by
writing the lists as declared by the YAML source.

Change-Id: Ie3f3de932a7b13b1dfd889f628790729b6651f44