multiversion: Let buildDBLists.php expand expression dblists
- Remove performance exemption for 'group1' in DBlistTest.

group1.dblist was the last remaining "expression" dblist used in
production, as a regular dblist it is now suitable for production.

- Move expression from .dblist to .dbexpr files.

The expressions still exist for easy maintenance of these lists,
in a separate dbexpr file. This no longer clashes with any
dblist files, and is preferable over e.g. calling these
"group1-expression.dblist" because that would mean that they'd
be valid to use in 'foreachwikiindblist' and in 'expanddblist'.
Which isn't bad per se, but it's better to only have one way
to do the same thing.

For simplicity sake, this change expands all (four) expression
dblists, not just group1. Specifically, it expands:
- open
- group1
- group2
- echo

One benefit this brings is that it will now be easy to review in
Git when the effective contents of a dblist changed. And when
implicit changes elsewhere have knock-on effects elsewhere, those
will be visible in CI and code review.

- BONUS: Remove support for expression expansion from readDbListFile()

Now that all lists are real dblists, this function no longer
has to be capable of expanding expressions.

Note that this doesn't affect command-line use cases such as
'foreachwikiindblist' and 'expanddblist', as their argument is
themselves considered a db expression as well. E.g. it wasn't
just possible to use "foreachwikiindblist echo" and have it
expand to the list of echo wikis, it was also possible to do
thins like "foreachwikiindblist echo - private", and that remains
possible.

Bug: T169821
Change-Id: I403a9890a98d47b8753db371c36f9b06e717177f