wgConf: Assign $wgLocalDatabases normally instead of by-ref
This by-ref has been here since at least the initial 2012 Git commit
of the public wmf-config repo (015f5b7131e, not in Gerrit).

I've audited all the search results for `(wg|'|")LocalDatabases`
and did not find any mutation of it within any wmf-deployed code.

I assume therefore that it was an memory perf optimisation from
the PHP 4 time – before there was copy-on-write in PHP.

Bug: T169821
Change-Id: Ie9002d9095eecef0371b70860463a6760c0e6d63