wgConf: Remove unused 'fullLoadCallback' property assignment
This hasn't been used in a long time. This property can be used
to make the SiteConfiguration lazy-load part of its settings array.
Certainly an interesting feature we may want to consider using
at some point, but we're not using it right now.

That is, wmf-config never calls $wgConf->loadFullData(), which
would call the function named in 'fullLoadCallback'.

That function is wmfLoadInitialiseSettings() which we already call
directly when needed in CommonSettings.php.

Bug: T169821
Change-Id: Ifc1511009fdbe1346c272753fcffe13d95f683fc