Make CSP enforce on beta
We're still probably a little ways away before doing this on
production, but lets start getting people used to this and
ensure they are testing with this in mind.

Expected breakage: Certain browser extensions that try to run
within the context of the page, may be broken by this change.
(The suggested solution is people make better browser extensions).
wikiplus-app.com will be blocked by this (intentional). Kartographer using
external map sources may break (T211971).

Bug: T244124
Change-Id: I9265315d9c9d11292c6d12887592351b64b264f4