Add prod domains to beta CSP policy to allow easier gadget testing
Some people seem to like to directly import their gadgets from prod
wikis in order to test them on beta. For now lets not break that.
If we want to enforce that nothing in beta depends on prod, that's
a discussion we can have later.

Note: This also unbreaks certain central notice things which
seemed to be calling from beta->prod, as well as some code looking
up Schema pages on meta. They probably shouldn't be working like
that on beta.

I had thought that prod domains would automatically be merged as the
setting didn't start with a "-", but that seemed to not happen.

Change-Id: I5d29a1df160e69fab80265682339a6940f21e027