cirrus: Increase commonswiki near match weight
Since we switched to BM25 and dropped the coord factor queries with a
lot of words may have greater scores than a perfect full title match.
Increase the weight to push near matches to the top. There is
relatively little risk as near matching is quite strict and will only
boost results with a almost the same title.

Bug: T245642
Change-Id: I69ba05eb40eb59b5b10a5df812eb9ab47aea5202