Don’t check constraints on two property qualifiers
The “is a list of” [1] and “category contains” [2] properties on
Wikidata are similar to the various “example” properties, in that any
other property can be used in the qualifiers of their statements, even
when that property is normally not supposed to be used as a qualifier.
Add them to the corresponding WikibaseQualityConstraints setting – and
while we’re at it, also split that setting across multiple lines, since
it’s getting rather long.

[1]: https://www.wikidata.org/wiki/Property:P360
[2]: https://www.wikidata.org/wiki/Property:P4224

Bug: T235292
Change-Id: If51303f0892686e2da8ea3dd0c818da8c914df90