MachineVision: Add MachineVisionWithholdImageList config
Adds a new variable definition to the production config containing a
list of Wikidata IDs indicating that an image should be withheld from
Special:SuggestedTags.

The earlier (slightly shorter) version is removed from the Beta config.

N.B. Items appearing on this list should not also appear in
$wgMachineVisionWikidataIdBlacklist because currently that list is used
for filtering labels before images are evaluated for withholding based
on this list. We don't want to filter items that could indicate that an
image should be withheld. The extension code will be updated soon so
that the blocklist filtering is applied after evaluation for
withholding, so that items can appear on both lists.

Bug: T249939
Change-Id: I600c22767dd6a5aef5d3e97f48a8c4eeb28cdabb