cirrus: redirect more_like to codfw to rebuild query cache
To rebuild the more_like query cache we need the capacity to serve
up to 900 qps of more_like traffic. Accomplish this by shifting the
traffic to the idle cluster in codfw and increasing the pool
counter limits.

Pool counter limits are increased from those used when we did
the same action in feb 2020 as traffic has increased ~20% since
then.

Change-Id: I33f46337a348069283a6561bae57812a4607dba2