Turn off $wgResourceLoaderUseObjectCacheForDeps in Beta and testwikis
This was enabled as part of potential expedited roll-out to unblock
the train (T247028), but we're instead going to keep the model as-is
for now and try to fix its issues instead. In part also because it
looks like the data, while relatively small, might be too big to
fit comfortably into the current Redis mainstash configuration.

This effectively reverts 537d0059f and af1b9c2dd.

Change-Id: Ic005093778d853acb009f50b765c121ad9b8d583