logging: Omit 'unique_id' from WebProcessor mixin
Inline its logic in our custom processor, and omit 'unique_id'
from it. I've filed a ticket upstream for making it possible
to omit this through a class option. It actually has a class
option already to specify the subset of fields you want to
include, but 'unique_id' just happens to bypass that filter
currently.

Bug: T253677
Change-Id: If991929c84ff69ee501c3075bf67664e0a919cfb