profiler: Add PDO driver for XHGui and enable on Beta Cluster
This adds a couple of classes based on the latest version of the XHGui
Collector for sending XHGui profiles to MariaDB, and updates the
profile callback to save using that destination if it is configured.

We don't yet have a database set up to receive these in production, so
right now this patch should only affect the labs realm.

Bug: T180761
Change-Id: I7721f4018b07dac3181b026f376cd6b776a38c25