Remove ContributionTracking extension
This extension was deployed long ago to redirect straight from
fundraising pages on foundationwiki and other wikis to our PayPal
pages.

Then we moved that logic to DonationInterface, but since
DonationInterface is deployed on the main cluster for its messages
(used in banners) and DonationInterface used to depend on
ContributionTracking to manage a database connection, we left
ContributionTracking deployed on the main cluster.

With commit If1c8fdab21c9029ea8132d , we remove the last need for
ContributionTracking as a pre-requisite for DonationInterface and
can uninstall ContributionTracking from the main cluster.

Change-Id: I5d1c0b2e5acb947ed69255a843cb6b6f09870e22