profiler: Fix undefined $wmgXhguiDBuser
Follows-up 40c56a76b. These are not available in PhpAutoPrepend
as none of MediaWiki and wmf-config are loaded yet at this point.
It also can't be loaded on-demand here as PrivateSettings builds
on top of MW defaults and the rest of wmf-config, which PHP
has not yet initalised at this point.

Evaluate it late instead, and accept that if MediaWiki fails
fatally during the first init phase, we cannot capture a profile
of that.

Bug: T180761
Change-Id: If7e1613cbcf8a0e920e8b342cf4719a1780334c5