logging: Fold Parsoid into type:mediawiki, add 'servergroup:' instead
=== Why

Error logs from all application servers is under type:mediawiki,
even for maintenance and snapshot CLI, and jobrunners, but,
not Parsoid.

This meant that if an error was caused by a Scap deploy changing
wmf-config or deploying mediawiki/parsoid code that affects
Parsoid, it would not be detected by Scap's Logstash checks.

It would also not be spotted by humans checking mediawiki-errors
in Kibana, and also not be spotted by Icinga alerts monitoring
Logstash levels.

To avoid incorrect risk calculations by deployers, fold it back
into type:mediawiki.

=== History

This essentially reverts I41b4b1820 (1a09e2a5a6), which made
sense at the time as Parsoid was still deployed out-of-bound.
Although even then some of the above did apply as well.

For the continued convenience of Logstash filters and dedicated
dashboards/alerts for Parsoid, a new key is added, 'servergroup'
which holds the value "parsoid" for the Parsoid servers.

This also has the added benefit of allowing MediaWiki's main
monitoring dashboards to be given a breakdown chart by servergroup
which can help a lot in better understanding the nature of
a particular incident. For example, it is difficult to tell right
now whether an issue solely affects API servers as the hostnames
all look alike, and the URLs differ.

Bug: T255627
Change-Id: I4c519f88c613fcec8f15e1a204f90943c7c10240