EventBus: Emit kafka purges for everything
This enables EventBus on wikitech and limits it to only
emitting purge events. Additionlly, login/vote wikis that
are special cased have EVENT_PURGE enabled as well. The purge
event is not yet marked as EVENT_PURGE in EventBus, this needs
to be added separately. Ence purges are marked as EVENT_PURGE,
we can switch wikitech and login/vote wikis to kafka purges,
concluding the transition.

Bug: T250781
Change-Id: I8d0c0b20ef3c43d4d94b40b1ec673520d6e5ef2b