multiversion: Fix 'closed-labs' reading as 'closed' for static config
Follows-up Idefc7cde70c489646101234, which changed the
`$dblist = MWWikiversions::readDbListFile()` call that populates
the wikiTags for static config to no longer use a realm-specific
file (at the time that code lived in CommonSettings.php).

This meant that eswiki was no longer closed in labs, and aawiki
was closed in beta.

By reading flow-labs and closed-labs as replacements it fixes
a number of things:

* closed-labs.dblist is once again independent of closed.dblist.
* Beta eswiki is read-only closed, again.
* Beta deploymentwiki is read-only closed (per Ide36b3fd32c98).
* maintenance scripts looping over 'closed-labs' will once again
  be inclusive instead of either missing all the closed ones or
  only getting the labs-only closed ones.
* The config hack for wmgUseFlow is no longer needed.
* There is no longer potential for other stuff relating to 'flow'
  in wmf-config being missed in beta or getting out of sync.
* maintenance scripts looping over flow-labs can assume it to be
  complete.

For later, we should see if "flow_only_labs" is still needed.
It has a FIXME to get rid of it, one of the cronjobs might need
updating, but it looks ripe for removal!

Bug: T109157
Bug: T198673
Bug: T115584
Change-Id: I64e21f8bf39e837130cfdd98d11e6b1893635331