labs: Set 'test.wikipedia.*' as canonical for Beta Cluster testwiki
The testwiki is configured as a Wikipedia-family project, with as canonical
test.wikipedia.org in prod..

Due to the weird and wonderful magic that is Multiversion, it was in fact
already accessible both through https://test.wikipedia.beta.wmflabs.org
and https://test.wikimedia.beta.wmflabs.org.

The former works because *.wikipedia is mapped both in Apache,
and in Multiversion with nothing stopping it from matching testwiki.

Once this is used as the canonical address, the Puppetised Beta-only
vhost for 'test.wikimedia.*' can be removed (see I135f54d7d27997922).

Bug: T99156
Change-Id: I56eb4a802940af1992ad9d33e392a1f9db6721aa