Scale largest shards to be closer to 30GB
This change was made as a result of a warning for >50 GB shard size on
commonswiki_file, and a review of other shards nearing the limit via a
quick glance at `curl -s -k
'https://localhost:9243/_cat/shards?s=store'`
identified enwiki_content and viwiki_general as also nearing the 50 GB
threshold.

We're hoping to buy enough time until we have 10G networking on the
entire cluster, but if not this will still buy us a lot of time.

Bug: T256928
Change-Id: Iccf74159e4acc6311d1de7bb95f17b06165783c4