Make testcommonswiki a testwikidata client
Apparently this group has been missing from the wiki since creation?
I'm not totally sure about the impact to testcommons of changing this now,
but this is probably the desired state?

This might explain why I can add a link to testcommonswiki from testwikidata
as of I825e1328ae025c6844869033fea2c4a346fdd183 as the config is not loaded
on test commons.

After a quick skim through IS.php this also will be causing other bad things,
for example the shared cache for production will be getting poluted from things
test commons is doing currently, as this group decides on that key prefix...

Bug: T257266
Change-Id: I46f589bd1b0cba954e0b497625a5d0be3e65eafa