Drop main page special casing on all projects
The grace period has now been completed.

Note, this will lead to minor breakage to the main page
of at most 61 projects, all
of which are low traffic, that have proved unresponsive
to the deprecation notice. All these projects have an audience of
less than 0.5 million hits (across all pages) a day.

It's expected that these will be fixed retroactively and
I will assist with those as they come up.

Bug: T32405
Bug: T254287
Change-Id: I7912710d5b6bb5108179369a82d08216acc6d22d