Wikibase: Split localEntitySourceName config for repo and client
This allows for different configuration to be defined between
repo and client, which is needed on commons.

This is analyzed in T254315#6304916 (and the comments just before)
and was the cause of the regression filed as T257266 when trying to
use entitysource based config for commons in production.

Beta and test commons sites are actually currently broken, but in
a hard to see way.

Setting this value from the old wmg values is left in place so that
this config can slowly be rolled out across beta -> test -> production.
Then the old wmg values will be removed.

Bug: T254315
Change-Id: Ia604bb2f21dcaa96cbeab5dacb8af0c12fb63449