wgEventLoggingSchemas - update bugged schema revisions (group0)
NavigationTiming and ResouceTiming are sending data that is
not possible to import into Hive. The schemas have been updated
with better validation rules, but we need a code deploy to make
EventLogging use the new schema revisions.  Override the schema
revisions used here in MW config until the owner code is updated.

This is group0 only, we will test and make sure it works and then
proceed with all wikis.

Bug: T254606
Change-Id: I698f55ad9e64287c29885b24bd11a6a394085547