Correct the wrong workmark and tagline for Chinese Wikimedia Project
Remove the current workmark and tagline for zh, because zh-classic is
not the same as zh, and its not possible having all Wikimedia Project using the same wordmark. Same thing goes for tagline.

Instead, adding the correct workmark and tagline for zhwiki with correct
width and height.

Note: Both the Simplified and Tranditional Chinese version of the
wordmark and tagline has been uploaded. However, only the Tranditional
Version is set in the InitialSetting.php as Chinese Wikipedia will do
the Simplified/Tranditional variant change via Chinese Wikipedia's own Common.css.

Bug: T260908
Bug: T258552
Change-Id: Ifc74e1f8498ba827bb1a5c42bea105c4a86ed273