wgEventStreams - Set canary_events_enabled: true for eventgate test streams
and eventlogging_Test

We will use this stream config setting to determine which streams we should
produce canary events into. Eventually, we'd like all streams to have canary events,
but we'll need to coordinate with consumers of canary event streams before we
do so.

To test this functionality we first enable this only for eventgate test streams.

Also remove a now unused test.event stream.

Bug: T251609
Change-Id: I768e5587796ebb3a0b9fa7dd794670346d4e6d7d