Remove unused $wgExtraLanguageNames['qqq'] assignment
I think this was dead code since change I7ecf369b18 (commit 1b6b56ea09),
which added a blanket `$wgExtraLanguageNames = $wmgExtraLanguageNames;`
override, removing the 'qqq' entry. Presumably, this was the cause of
bug 46064 (T48064), which was then fixed in the Translate extension
(change Ie7fb561891), so that it is no longer necessary to add 'qqq' to
the extra language names in the wiki configuration.

Bug: T263441
Change-Id: I02f577d339fdb53a5c14dc9076986b3c58d41cfc