Remove commonswiki from sister search sidebar
Commonswiki is one of our largest indices. Try and reduce the size
of the in-memory working set by removing it from the sidebar. In a
previous analysis sister search performs an order of magnitude more
queries against commonswiki than commonswiki itself.

Not 100% sure this will help, but we should be able to tell from
post-deployment graphs.

Bug: T264053
Change-Id: I5f960f8d01ea94138f0b24492d9ddeeb53802317