cirrus: Increase more_like cache from one to three days
Analysis of query logs shows that more_like typically returns the same
results over windows of several days. The search clusters additionally
need to shed some load to keep up with recent expansions in data size.
Shed some of our more expensive queires by caching the results
longer.

These are *also* edge cached, but we decided against changing smaxage
in the api requests as that is a cache-busting maneuver that can be
managed when things are running well, but isn't ideal when trying
to shed load.

Bug: T264053
Change-Id: I7860401dd796fd8f45b3b4fe73ee7a805b10a805