Enable parsoid on api_appserver
The MW core REST /html endpoint will be using Parsoid directly,
without going to RESTBase. Some of the glue code between core
and Parsoid still resides in the parsoid extension, see T265518,
so in order for the API to keep working, we need to enable Parsoid
extension on api_appserver cluster.

This enables Parsoid extension on api_appserver, but does not enable
Parsoid REST API.

Bug: T265954
Depends-On: I69d171bad2a7d1a96537444869b66524171997ae
Change-Id: I49d1dd772025203b4560487a57ba888c7870d539