cirrus: Hardcode more_like to codfw cirrus cluster
This is a temporary measure that we are putting in place before the
yearly datacenter switchback. Currently the active cluster for
cirrussearch is codfw, but it will be switched back to eqiad in a matter
of days.

Since eqiad has been offline for some time (per the switchover), we will
experience a flood of cirrus cache misses when we switchback. As such,
by hardcoding more_like queries to route to codfw's cirrus cluster, we
will be able to lean on cirrus elastic instances to populate
the cache of eqiad mediawiki app instance.

We'll leave this mitigation in place for 24 hours following the
switchback - which is the cache length. At that point we'll revert this
and let eqiad serve more_like queries normally.

Change-Id: Id7374056d5c4adc8d20ba2f2a5e6bf24f48507c1