Special docroot for thankyou.wp.org (and donate)
Thankyou.wikipedia.org should be excluded from opening in apps, which
by default launch for all the *.wikipedia.org subdomains.

This adds a special apple-app-site-association file for thankyouwiki
which prevents it opening in iOS. For Android apps, the change was
made in the code that handles activity launches.

This will need a matching patch in the main cluster puppet repo.

Bug: T259312
Change-Id: I2aefcf167f273d357ac93a7fa57447d2ed346d34