Enable canary events for recommendation create
We are setting up the processing for these events in the
analytics network. For those processes to work they need
a guarantee of at least 1 event generated every hour, so
analytics jobs have a clear marker of when an hour with no
events has been imported to hdfs.

Enable canary events, which was designed for exactly this purpose.

Bug: T262226
Change-Id: Ia1e9b6fec8551ac487804264ac55b9d7c6cdf185