logos: Redo how variants work
Rather than overwrite the main "commons" value when you want to add a
variant, they're now specified in a "variants" subitem. An additional
`--variant` option has been added to the "update" subcommand to
download/compress variant logos.

The previous "variant" key has been renamed to "selected" as it's the
name of the logo that has been selected to be in logos.php.

The primary motivation for this redo is because language converter wikis
need to always have a variant logo. See the next patch for how this
works in practice for zhwiki.

Bug: T98640
Change-Id: I12f68efc6c16b20e79e430520848cf14c58fddb4