fatal-error.php: Add support for from=destruct and from=shutdown
Uncaught exceptions are generally reported to Logstash by MW and not
the php-wmerrors extension, unless they happen during the shutduwn
stage from a shutdown or destructor callback.

That case was not previously testable with this script, but is not
possible with from=destruct and from=shutdown.

This covers a gap in testing for errors that happen after the stack
unwinds. The "postsend" case from before is more about MW-specific
logic and doesn't really have significant for the PHP engine state and
error reporting pipeline since it's still part of the main stack at
that point.

Bug: T275075
Change-Id: I9a050c4bb0c62953ee77568bd7472875dbcfb1fe