[stage 1] Enable WVUI search by default to logged-in modern Vector users except on pilot wikis
According to T249297, stage 1 deployment should enable WVUI search on
modern Vector to all logged in users except on pilot wikis.

This patch uses the `wgVectorSearchTreatmentABTest` flag to enable the
A/B test by default except to the pilot wikis (`desktop-improvements`
group). As a result, half of logged-in modern vector users on non-pilot
wikis will see the WVUI search and the other half will see legacy search
- a condition that has been approved by our product manager. Also,
because of code in Vector [1], note that the
`wgVectorSearchTreatmentABTest` flag can enable WVUI search even if the
`wgVectorUseWvuiSearch` flag is `false` which it currently defaults to.

Because `testwiki` and `test2wiki` are enabled through the
`wgVectorUseWvuiSearch` flag, these wikis will continue showing WVUI
search to logged in and anon users.

[1] https://github.com/wikimedia/Vector/blob/31a9e6feff5d926b44affb5303ce292bee8ba389/includes/FeatureManagement/Requirements/WvuiSearchTreatmentRequirement.php#L74-L83

Bug: T249297
Change-Id: Ica3414f05c6b4294c1ca2131e2023faed5a7147c