wmf-config/CommonSettings.php: Add WMF_MAINTENANCE_OFFLINE handling
There are some maintenance scripts that we want to be able to run
offline (meaning without accessing external databases).  For example,
to build production-ready mediawiki images we need to be able to run
the mergeMessageFileList.php and rebuildLocalisationCache.php
maintenance scripts offline.

This commit makes wmf-config/CommonSettings.php check for the
existence of the WMF_MAINTENANCE_OFFLINE environment variable.  When
set, CommonSettings.php will mock various settings to allow remaining
config to be loaded without generating warnings or trying to access
etcd.

Bug: T238436
Bug: T237148
Bug: T260827
Change-Id: Ib4bba62f290cb16dcf43ba7755b2eeedfa7e5901