wikitech: enable BetaFeatures
BetaFeatures is disabled on Wikitech, I notably lack the VisualEditor
wikitext editor.

The extension has been disabled in 2014 by 1f50426, probably to simplify
an hardware migration though the commit message lack any justification
and does not point to any task.

The extension requires the database table `betafeatures_user_counts`
which is present on the `labswiki` database and does match the extension
schema.

Bug: T125941
Change-Id: Idd19a0c6f5e87f5be08f95cd1a2f824cf7f5f742