pipeline: add building the webserver image
We need to have all the endpoints and the docroot in the webserver image.

Ideally in a future we'll be able to split the repository between one where
we just have the endpoints and the static assets, and one where we have
the php code/libraries. For now, just resort to a long list of excludes.

Change-Id: I987307d54e436aa8ac2d245c30c7ac569c15bacb