Enable DiscussionTools' beta features on almost all remaining projects
Leaving all of the redundant values for now: the Phabricator task
references might still be useful to keep track of which communities
requested it to be enabled earlier.

On a few wikis, the reply tool is already enabled for everyone (not
part of a beta feature), that is not changed.

German Wikipedia (dewiki) is not included: we're planning to discuss
it with that community separately (T275254).

Bug: T276498
Change-Id: I3f662f0cc57ede4bc12513b214ad71c2a10d708e