Define confirmed group in MediaWikiServices hook
As explained in T275334#6855879, extension functions can no longer
manage permissions. @daniel proposed to use MediaWikiServices as
a hacky way to do the same.

Until the Platform Engineering figures out how to solve T275334 in
the long term, let's define this group in MediaWikiServices hook,
so it can be actually always used.

Bug: T275334
Bug: T277704
Change-Id: I5a7f9da835d07d5858b38778015cc6fa9be22d9b