Disable legacy javascript global variables in ruwiki
We are disabling in the last wikis one by one,
zhwiki cleaned up of any error indicating that can move on
to the next wiki with low risk of triggering an alert:
https://grafana.wikimedia.org/d/000000566/overview?viewPanel=16&orgId=1&from=now-12h&to=now

I did a preliminary clean up of all wikis in the past months (including ruwiki),
we are not going there blind but there might be edge cases or stuff my script didn't catch.
That's why we disable them one by one.

Bug: T72470
Change-Id: I20e58df6a97c12158686b2dd90aa89fabe9a1426